During vertex kinfit, propagate measured params to POCA to kinfit position.
The choice of a point on the track is arbitrary. We choose POCA to the beamline.
However, the kinfit is also free to choose a different point.
This may throw off the chisq and pulls.
So, after each iteration, choose a NEW measured point on the measured trajectory.
This point is the point nearest the chosen kinfit point.
Update the relevant matrices with this choice.